Reimplemented EBeam library
