+++ simphony/DeviceLibrary/sparams/ebeam_dc_halfring_te1550/te_ebeam_dc_halfring_straight_gap=30nm_radius=3um_width=520nm_thickness=210nm_CoupleLength=0um.dat
