+++ simphony/DeviceLibrary/sparams/ebeam_gc_te1550/GC_TE1550_thickness=220 deltaw=0.txt
