+++ simphony/DeviceLibrary/sparams/ebeam_terminator_te1550/nanotaper_w1=500,w2=60,L=10_TE.sparam
