+++ simphony/DeviceLibrary/sparams/ebeam_wg_integral_1550/WaveGuideTETMStrip,w=500,h=220.txt
