+++ simphony/DeviceLibrary/sparams/ebeam_wg_integral_1550/straightCoeffs.npy
