--- simphony/elements/__init__.py
@@ -1,10 +1,0 @@
-__all__ = [
-    'ebeam_bdc_te1550',
-    'ebeam_dc_halfring_te1550',
-    'ebeam_gc_te1550',
-    'ebeam_terminator_te1550',
-    'ebeam_wg_integral_1550',
-    'ebeam_y_1550',
-]
-
-from . import *
