--- simphony/elements/ebeam_dc_halfring_te1550/tests.py
@@ -1,3 +1,0 @@
-
-
-# Create your tests here.
