--- simphony/elements/ebeam_y_1550/tests.py
@@ -1,3 +1,0 @@
-
-
-# Create your tests here.
